% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_school.R
\name{check_school}
\alias{check_school}
\title{Summarize the match between a school spreadsheet and a template}
\usage{
check_school(.school, .regions, .template)
}
\arguments{
\item{.school}{the spreadsheet that represents a school}

\item{.regions}{a list of regions of the spreadsheets that should be compared}

\item{.template}{the canonical layout that the school should have followed}
}
\value{
a list with two kinds of elements:
\itemize{
\item For each name in \code{.regions}, a count of the number of mismatches.
\item \strong{messages} a named character vector of mismatch details
}
}
\description{
Summarize the match between a school spreadsheet and a template
}
