% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sheet.R
\name{check_sheet}
\alias{check_sheet}
\title{For one worksheet layout, compare every school to a template}
\usage{
check_sheet(.schools, .regions, .template)
}
\arguments{
\item{.schools}{a list of worksheets named by the school they come from}

\item{.regions}{a named list of regions within the template to examine}

\item{.template}{the template that serves as the standard}
}
\value{
a tibble with the following fields
\itemize{
\item \strong{School} the label of the school
\item A count of the number of mismatches for each element of \code{.regions}
\item \strong{messages} a list of the mismatches
}
}
\description{
For one worksheet layout, compare every school to a template
}
