% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_template.R
\name{check_template}
\alias{check_template}
\title{Compare a region of a submitted worksheet to its template}
\usage{
check_template(.dims, .template, .submission)
}
\arguments{
\item{.dims}{A list with elements \code{rows} and \code{cols}}

\item{.template}{A tibble of text-valued cells}

\item{.submission}{A tibble of text-value cells to compare to \code{.template}}
}
\value{
A \code{waldo::waldo_compare} that describes each mismatch between inputs.
}
\description{
Compare a region of a submitted worksheet to its template
}
