% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COLLEGE_SCORECARD_SCHEMA}
\alias{COLLEGE_SCORECARD_SCHEMA}
\title{Metadata for reading College Scorecard data for HERA dashboards}
\format{
a tibble with eight columns:
\itemize{
\item \verb{API data type} \emph{character} - usually either "float" or "integer".
\item \code{Variable} \emph{character} - The name of the variable in the US Ed db.
\item \code{Measure} \emph{character} - The actual quantity being measured.
\item \verb{Degree Duration} \emph{character} - Are we talking about a 2y or 4y cohort
\item \verb{Demographic Category} \emph{character} - Race, financial aid, sex, etc.
\item \verb{Demographic Detail} \emph{character} - The specific level within the category.
\item \verb{Enrollment Status} \emph{character} Full- or Part-time.
\item \code{Specification} \emph{character} - depends upon \verb{API data type}.
}
}
\usage{
COLLEGE_SCORECARD_SCHEMA
}
\description{
Metadata for reading College Scorecard data for HERA dashboards
}
\keyword{datasets}
