% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardify_grads_times_credits.R
\name{dashboardify_grads_times_credits}
\alias{dashboardify_grads_times_credits}
\title{Convert Higher Expectations' data layout to HERA's dashboard layout}
\usage{
dashboardify_grads_times_credits(.x)
}
\arguments{
\item{.x}{\emph{tibble} - a data frame produced by \code{read_grads_times_credits}.}
}
\value{
a data frame in the HERA dashboard format.
}
\description{
Convert Higher Expectations' data layout to HERA's dashboard layout
}
