% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_enrollment_status_in_population.R
\name{detect_enrollment_status_in_population}
\alias{detect_enrollment_status_in_population}
\title{Predicate for finding rows that contain full- or part-time status}
\usage{
detect_enrollment_status_in_population(.v)
}
\arguments{
\item{.v}{A vector of population names}
}
\value{
a logical vector that is true where the word "time" appears in \code{.v}
}
\description{
Predicate for finding rows that contain full- or part-time status
}
\examples{
detect_enrollment_status_in_population(c("time", "nope", "Time", "full-time"))
}
