% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment_status_from_population.R
\name{enrollment_status_from_population}
\alias{enrollment_status_from_population}
\title{Create a column of enrollment statuses from the population column}
\usage{
enrollment_status_from_population(.v)
}
\arguments{
\item{.v}{a character vector of grouped population characteristics}
}
\value{
a character vector of enrollment statuses
}
\description{
Create a column of enrollment statuses from the population column
}
\examples{
enrollment_status_from_population(
 c("Full-Time (at entry)","TOTAL Students", "Race", "Gender", "Age")
)
}
