% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_and_name_files.R
\name{expand_and_name_files}
\alias{expand_and_name_files}
\title{Zip together file names and factor labels, optionally rooted at a common path}
\usage{
expand_and_name_files(.files, .path = NULL, .labels = NULL)
}
\arguments{
\item{.files}{a vector of file names}

\item{.path}{optional, a path to prepend to the file names}

\item{.labels}{optional, names for each file}
}
\value{
A named character vector of full paths
}
\description{
Zip together file names and factor labels, optionally rooted at a common path
}
