% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_down_when_blank.R
\name{fill_down_when_blank}
\alias{fill_down_when_blank}
\title{Turn human-readable blank-riddled columns into actual data}
\usage{
fill_down_when_blank(.v)
}
\arguments{
\item{.v}{a column of data with \code{NA}'s in empty cells}
}
\value{
a column of data with \code{NA}s replaced by values above them.
}
\description{
Turn human-readable blank-riddled columns into actual data
}
\examples{
v <- rep(x = c("A", NA, "B", NA, "C", NA),
         times = c(1, 3, 1, 4, 1, 2))
fill_down_when_blank(v)
}
