% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_matching_files.R
\name{find_matching_files}
\alias{find_matching_files}
\title{Search a vector of file names for a pattern}
\usage{
find_matching_files(.pattern, .files)
}
\arguments{
\item{.pattern}{string - A regular expression to search for}

\item{.files}{string - A vector of file names}
}
\value{
a single match, \code{NA}, or an error
}
\description{
Search a vector of file names for a pattern
}
\examples{
find_matching_files("f", c("foo", "bar", "baz", "~$foo"))
}
