% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HERA_SCHOOLS}
\alias{HERA_SCHOOLS}
\title{Identifying information for finding HERA schools in the College Scorecard}
\format{
A tibble with 6 columns:
\itemize{
\item \code{UNITID} \emph{integer} - College Scorecard code for one specific campus/location.
\item \code{OPEID6} \emph{integer} - A code for one IHD, which may have multiple campuses.
\item \code{Name} \emph{character} - The name of the Institution of Higher Education.
\item \code{Campus} \emph{character} - The name for the campus, optional
\item \code{Search} \emph{character} - A search term for finding files relevant to the school.
\item \code{Institution} \emph{character} - The school's name in HERA's Tableau dashboard.
}
}
\source{
\href{https://collegescorecard.ed.gov/data/}{College Scorecard}
}
\usage{
HERA_SCHOOLS
}
\description{
The U.S. Department of Education maintains a database of demographic and
performance measures for participating institutions of higher education.
Southeastern Wisconsin has a confederation of such schools, the Higher
Education Regional Alliance.
}
\keyword{datasets}
