% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_batch_read_tables.R
\name{ipeds_batch_read_tables}
\alias{ipeds_batch_read_tables}
\title{Title}
\usage{
ipeds_batch_read_tables(
  schema,
  survey_code,
  table_code,
  start_year,
  finish_year = NULL,
  data_path = "."
)
}
\arguments{
\item{schema}{The table that describes how the source data tables are laid out.}

\item{survey_code}{The IPEDs prefix that designates related tables.}

\item{table_code}{The IPEDS suffix that designates a specific table.}

\item{start_year}{The first year to look up.}

\item{finish_year}{The last year to look up.}

\item{data_path}{The path to where the data files are stored.}
}
\value{
A (probably huge) tibble with data from many schools. It will include
a column "Year" in addition to the columns dictated by \code{schema}.
}
\description{
Title
}
\seealso{
\code{\link[=ipeds_zipped_csv]{ipeds_zipped_csv()}} the function that opens compressed archives.

\code{\link[=ipeds_read_with_schema]{ipeds_read_with_schema()}} the function that parses each file.
}
