% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.R
\name{ipeds_data_file}
\alias{ipeds_data_file}
\alias{ipeds_data_zip}
\alias{ipeds_data_csv}
\title{The file names for IPEDS data from specific years on specific topics}
\usage{
ipeds_data_zip(.year, .survey_code, .table_code)

ipeds_data_csv(.year, .survey_code, .table_code)
}
\arguments{
\item{.year}{integer - A four-digit integer value for a year.}

\item{.survey_code}{character - One of the IPEDS prefixes that designate common topics.}

\item{.table_code}{character - One of the IPEDS suffixes that specify tables within topics.}
}
\value{
The file name as a single string.
}
\description{
These functions all glue together the inputs and a file extension. They all
change the case of the text inputs to match the arbitrary but consistent
formatting style of the type of file.
}
\section{\code{ipeds_data_zip()}}{


Compressed files from the website or saved somewhere by you.
}

\section{\code{ipeds_data_csv}}{


The contents of a compressed file.
}

\examples{
ipeds_data_zip(2002L, "ef", "d")
ipeds_data_csv(2002L, "ef", "d")
ipeds_data_zip(2012, "c", "c_")
ipeds_data_csv(2012, "c", "c_")
}
