% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.R
\docType{data}
\name{IPEDS_DATA_URL}
\alias{IPEDS_DATA_URL}
\title{The root path where annual data files can be retrieved from IPEDS}
\format{
a URL as a string
}
\source{
\href{https://nces.ed.gov/ipeds/datacenter/data/}{IPEDS}
}
\usage{
IPEDS_DATA_URL
}
\description{
The root path where annual data files can be retrieved from IPEDS
}
\keyword{datasets}
