% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.R
\name{ipeds_download_data}
\alias{ipeds_download_data}
\title{Fetch a zipped archive of data from IPEDs.}
\usage{
ipeds_download_data(.year, .survey_code, .table_code, .path)
}
\arguments{
\item{.year}{integer - A four-digit integer value for a year.}

\item{.survey_code}{character - One of the IPEDS prefixes that designate common topics.}

\item{.table_code}{character - One of the IPEDS suffixes that specify tables within topics.}

\item{.path}{character - The full path to where you want the .zip archive to be saved.}
}
\value{
The path to the downloaded file (invisibly).
}
\description{
Fetch a zipped archive of data from IPEDs.
}
\seealso{
\code{\link[=ipeds_data_zip]{ipeds_data_zip()}} correctly creates the file name you need

\code{\link[curl:curl_download]{curl::curl_download()}} fetches the file and saves it where you asked.
}
