% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_read_schema.R
\name{ipeds_read_schema}
\alias{ipeds_read_schema}
\title{Read one schema that describes a layout of IPEDS data.}
\usage{
ipeds_read_schema(.file, .sheet)
}
\arguments{
\item{.file}{character/connection - The spreadsheet with the schemas.}

\item{.sheet}{character - The specific sheet within the spreadsheet file.}
}
\value{
a tibble that describes how factor levels are coded
}
\description{
Read one schema that describes a layout of IPEDS data.
}
