% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_read_schema.R
\name{ipeds_read_with_schema}
\alias{ipeds_read_with_schema}
\title{Use a schema to parse a data table from IPEDS}
\usage{
ipeds_read_with_schema(.file, .schema, .nas = c("", "NA", "."))
}
\arguments{
\item{.file}{character/connection - The spreadsheet with the schemas.}

\item{.schema}{tibble - a tibble that describes how factor levels are coded.}

\item{.nas}{character - a vector of values to treat as missing.}
}
\value{
a tibble with tidy settings created with reference to \code{schema}
}
\description{
Use a schema to parse a data table from IPEDS
}
