% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_wrangle_completions.R
\name{ipeds_wrangle_completions}
\alias{ipeds_wrangle_completions}
\title{Wrangle imported IPEDS data about graduation into dashboardable long format}
\usage{
ipeds_wrangle_completions(.demographic, .completions, .schema)
}
\arguments{
\item{.demographic}{which demographic to wrangle}

\item{.completions}{the source tibble with lots of different demographics}

\item{.schema}{a tibble that describes how the demographic groups nest.}
}
\value{
a tibble of tidy graduation data
}
\description{
Wrangle imported IPEDS data about graduation into dashboardable long format
}
