% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_wrangle_retention.R
\name{ipeds_wrangle_retention}
\alias{ipeds_wrangle_retention}
\title{Wrangle imported IPEDS data about retention into dashboardable long format}
\usage{
ipeds_wrangle_retention(.retentions, .schema)
}
\arguments{
\item{.retentions}{a source tibble with lots of different demographics}

\item{.schema}{a tibble that describes how the demographic groups nest.}
}
\value{
a tibble of tidy retention data
}
\description{
Wrangle imported IPEDS data about retention into dashboardable long format
}
