% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.R
\name{ipeds_zipped_csv}
\alias{ipeds_zipped_csv}
\title{Open an IPEDS csv table from within a zipped file.}
\usage{
ipeds_zipped_csv(.year, .survey_code, .table_code, .path)
}
\arguments{
\item{.year}{integer - A four-digit integer value for a year.}

\item{.survey_code}{character - One of the IPEDS prefixes that designate common topics.}

\item{.table_code}{character - One of the IPEDS suffixes that specify tables within topics.}

\item{.path}{character - The full path to the folder where the .zip file is located.}
}
\value{
An open connection that can access the zipped data as if they weren't.
}
\description{
Open an IPEDS csv table from within a zipped file.
}
\seealso{
\code{\link[=ipeds_data_zip]{ipeds_data_zip()}}, \code{\link[=ipeds_data_csv]{ipeds_data_csv()}} Make file names with IPEDS style.

\code{\link[archive:archive_read]{archive::archive_read()}} The function that actually opens the archive.
}
