% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MATH_PATHWAY_NAMES}
\alias{MATH_PATHWAY_NAMES}
\title{Standard labels for the different possible Math Pathways}
\format{
a character vector
}
\usage{
MATH_PATHWAY_NAMES
}
\description{
Standard labels for the different possible Math Pathways
}
\keyword{datasets}
