% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mismatch_as_markdown.R
\name{mismatch_as_markdown}
\alias{mismatch_as_markdown}
\title{Print a \code{waldo_compare} report for inclusion in a Markdown document.}
\usage{
mismatch_as_markdown(.waldo_compare, .label, .heading)
}
\arguments{
\item{.waldo_compare}{\itemize{
\item output from calling waldo::compare
}}

\item{.label}{\itemize{
\item a text label for the comparison
}}

\item{.heading}{\itemize{
\item the heading prefix for the Markdown output
}}
}
\value{
nothing
}
\description{
Print a \code{waldo_compare} report for inclusion in a Markdown document.
}
