% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOMENTUM_MILESTONES}
\alias{MOMENTUM_MILESTONES}
\title{Map from short names for momentum milestone statuses to their prolix HERA labels.}
\format{
a tibble with four columns:
\itemize{
\item \code{Order} \emph{integer} The order that the milestone status appears in the template sheet.
\item \verb{Short Name} \emph{character} - A terse description of the milestone status
\item \code{Status} \emph{character} - Either "Completed" or "Exempt".
\item \verb{Long Name} \emph{character} - The full name of the measure from the template sheet.
}
}
\usage{
MOMENTUM_MILESTONES
}
\description{
Map from short names for momentum milestone statuses to their prolix HERA labels.
}
\keyword{datasets}
