% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_print_regions.R
\name{pretty_print_regions}
\alias{pretty_print_regions}
\title{Print a list of template regions as a Markdown pipe table}
\usage{
pretty_print_regions(.regions)
}
\arguments{
\item{.regions}{A list of lists of regions, named by worksheet}
}
\value{
Nothing, prints out the pipe table for pasting into documentation.
}
\description{
Print a list of template regions as a Markdown pipe table
}
