% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_a_sheet.R
\name{read_a_sheet}
\alias{read_a_sheet}
\title{Apply an importing function to one sheet from each of several spreadsheets}
\usage{
read_a_sheet(.files, .path, .schools, .function)
}
\arguments{
\item{.files}{A vector of file names}

\item{.path}{A common path (if any) where the files are located}

\item{.schools}{A vector of school labels, one per file}

\item{.function}{A sheet-reading function like \code{read_cc_or_tech_graduation}}
}
\value{
a tibble with outputs from
}
\description{
Apply an importing function to one sheet from each of several spreadsheets
}
