% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bachelors_graduation.R
\name{read_bachelors_graduation}
\alias{read_bachelors_graduation}
\title{Import data on graduation from programs that grant Bachelor's degrees.}
\usage{
read_bachelors_graduation(.file_name)
}
\arguments{
\item{.file_name}{the full path to the spreadsheet with data}
}
\description{
Import data on graduation from programs that grant Bachelor's degrees.
}
