% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cc_or_tech_graduation.R
\name{read_cc_or_tech_graduation}
\alias{read_cc_or_tech_graduation}
\title{Import data on graduation from programs designed to take less than 4 years.}
\usage{
read_cc_or_tech_graduation(.file_name)
}
\arguments{
\item{.file_name}{the full path to the spreadsheet with data}
}
\description{
Import data on graduation from programs designed to take less than 4 years.
}
