% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_glossary.R
\name{read_glossary}
\alias{read_glossary}
\title{Read in every sheet from a spreadsheet}
\usage{
read_glossary(.file_path)
}
\arguments{
\item{.file_path}{the full path to the excel workbook}
}
\value{
a list of data frames, one for each sheet in the workbook
}
\description{
Read in every sheet from a spreadsheet
}
