% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grads_times_credits.R
\name{read_grads_times_credits}
\alias{read_grads_times_credits}
\title{Import data on graduation times and credits earned across populations.}
\usage{
read_grads_times_credits(.file_name)
}
\arguments{
\item{.file_name}{the full path to the spreadsheet with data}
}
\description{
Import data on graduation times and credits earned across populations.
}
