% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_graduation_cc_or_tech.R
\name{read_graduation_cc_or_tech}
\alias{read_graduation_cc_or_tech}
\title{Import data on graduation from programs designed to take less than 4 years.}
\usage{
read_graduation_cc_or_tech(.file_name)
}
\arguments{
\item{.file_name}{the full path to the spreadsheet with data}
}
\value{
A tibble with seven fields:
\itemize{
\item \verb{Enrollment Status} \emph{character} - Either Full- or Part-time.
\item \code{Factor} \emph{character} - What type of demographic group this row describes.
\item \code{Level} \emph{character} - The specific demographic group that this row describes.
\item \verb{Completion Time} \emph{character} - Either 100\% or 150\%, which is relative to \verb{Program Length}.
\item \verb{Program Length} \emph{character} - Either "0y-2y" or "2y-4y".
\item \verb{First-time cohort} \emph{integer} - The number of students in this group in their first fall.
\item \code{Count} \emph{integer} - The number of students in the group who earned a credential.
}
}
\description{
Import data on graduation from programs designed to take less than 4 years.
}
