% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_momentum.R
\name{read_momentum}
\alias{read_momentum}
\title{Import data on student progress toward first-year goals.}
\usage{
read_momentum(.file_name)
}
\arguments{
\item{.file_name}{the full path to the spreadsheet with data}
}
\value{
A tibble with seven fields:
\itemize{
\item \verb{Enrollment Status} \emph{character} - Either Full- or Part-time.
\item \code{Factor} \emph{character} - What type of demographic group this row describes.
\item \code{Level} \emph{character} - The specific demographic group that this row describes.
\item \code{Milestone} \emph{character} - A specific first-year goal, such as completing a minimum number of credits.
\item \code{Method} \emph{character} - Whether the student reached the milestone by completing it or through exemption.
\item \verb{First-time cohort} \emph{integer} - The number of students in this group in their first fall.
\item \code{Count} \emph{integer} - The number of students in the group who achieved the milestone.
}
}
\description{
Import data on student progress toward first-year goals.
}
