% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_retention.R
\name{read_retention}
\alias{read_retention}
\title{Import data on student retention from their first fall to their second.}
\usage{
read_retention(.file_name)
}
\arguments{
\item{.file_name}{the full path to the spreadsheet with data}
}
\value{
A tibble with five fields:
\itemize{
\item \verb{Enrollment Status} \emph{character} - Either Full- or Part-time.
\item \verb{Demographic Group} \emph{character} - What type of demographic group this row describes.
\item \code{Detail} \emph{character} - The specific demographic group that this row describes.
\item \code{Cohort} \emph{integer} - The number of students in this group in their first fall.
\item \code{Count} \emph{integer} - The number of students in the group who returned the following fall.
}
}
\description{
Import data on student retention from their first fall to their second.
}
