% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_template.R
\name{read_template}
\alias{read_template}
\title{Read in the raw contents of a template spreadsheet}
\usage{
read_template(path_to_file)
}
\arguments{
\item{path_to_file}{\itemize{
\item string The full path to the template file
}}
}
\value{
a list of all-character tibbles named from sheet names
}
\description{
Read in the raw contents of a template spreadsheet
}
