% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relocate_for_dashboard.R
\name{relocate_for_dashboard}
\alias{relocate_for_dashboard}
\title{Set the order of a tibble to match what the HERA template expects.}
\usage{
relocate_for_dashboard(.x)
}
\arguments{
\item{.x}{a tibble of data from one of the \verb{wihera::read_*} functions.}
}
\value{
A tibble with at least 8 fields:
\itemize{
\item \verb{Source Tab} \emph{character} - The name of the sheet that the data came from.
\item \code{Degree/Certificate} \emph{character} - The pertinent credential track for the \code{Measure}.
\item \code{Measure} \emph{character} - The indicator that is quantified by \code{Outcome}.
\item \verb{Enrollment at entry} \emph{character} - Whether the students are full- or part-time.
\item \verb{Demographic Group} \emph{character} - A type of demographic group, like race or age.
\item \code{Detail} \emph{character} - A specific demographic group, like \code{Latine} or \code{18-24}.
\item \code{Cohort} \emph{integer} - If present, the denominator to use to convert \code{Outcome} to a percentage.
\item \code{Outcome} \emph{numeric} - A quantitative \code{Measure} of some aspect of the school's performance.
}
}
\description{
Set the order of a tibble to match what the HERA template expects.
}
