% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TEMPLATE_REGIONS_2023}
\alias{TEMPLATE_REGIONS_2023}
\title{Layout information for the data-entry template of 2023}
\format{
This list has the following layout: Sheet(Region(Rows, Columns)),
and the following values:\tabular{llrr}{
   Sheet \tab Region \tab Rows \tab Columns \cr
   Retention \tab fields \tab 1:5 \tab 1:3 \cr
    \tab labels \tab 6:41 \tab 1:1 \cr
   Momentum in First Year \tab fields \tab 1:5 \tab 1:10 \cr
    \tab labels \tab 6:41 \tab 1:1 \cr
   Math Pathways \tab fields \tab 1:6 \tab 1:9 \cr
    \tab labels \tab 7:11 \tab 1:1 \cr
   Graduation - Programs < 4 years \tab fields \tab 1:7 \tab 1:7 \cr
    \tab labels \tab 8:43 \tab 1:2 \cr
   Graduation - Bachelors degrees \tab fields \tab 1:5 \tab 2:5 \cr
    \tab labels \tab 6:41 \tab 2:2 \cr
   Graduates Time and Credits \tab fields \tab 1:5 \tab 1:5 \cr
    \tab labels \tab 6:149 \tab 1:2 \cr
}
}
\usage{
TEMPLATE_REGIONS_2023
}
\description{
Layout information for the data-entry template of 2023
}
\keyword{datasets}
