% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose_completion_time.R
\name{verbose_completion_time}
\alias{verbose_completion_time}
\title{Add extra verbiage to a percentage completion time.}
\usage{
verbose_completion_time(.completion_time)
}
\arguments{
\item{.completion_time}{\emph{character} - a percentage including a "\%".}
}
\value{
"Earned degree/certificate within {.completion time} time"
}
\description{
Add extra verbiage to a percentage completion time.
}
\examples{
verbose_completion_time(c("100\%", "150\%", "999\%", NA))
}
