% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_population_and_enrollment_status.R
\name{wrangle_population_and_enrollment_status}
\alias{wrangle_population_and_enrollment_status}
\title{The Population field includes sparse representation of Enrollment Status}
\usage{
wrangle_population_and_enrollment_status(.x)
}
\arguments{
\item{.x}{a data frame with a multi-valued HERA Population column}
}
\value{
A data with \code{Population} removed, and two extra columns:
\itemize{
\item \verb{Enrollment Status}: whether the students were full- or part-time.
\item \code{Factor}: Which kind of population the next variable denotes
\item \code{Level} : a specific kind of population, like a race or economic group
}
}
\description{
The Population field includes sparse representation of Enrollment Status
}
