% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardify_grads_times_credits.R
\name{dashboardify_grads_times_credits}
\alias{dashboardify_grads_times_credits}
\title{Convert Higher Expectations' data layout to HERA's dashboard layout}
\usage{
dashboardify_grads_times_credits(.x)
}
\arguments{
\item{.x}{\emph{tibble} - a data frame produced by \code{\link[=read_grads_times_credits]{read_grads_times_credits()}}.}
}
\value{
a data frame in the HERA dashboard format. It does not have the \code{Cohort} field.
\itemize{
\item \code{Degree/Certificate} \emph{character} - The pertinent credential track for the \code{Measure}.
\item \code{Measure} \emph{character} - The indicator that is quantified by \code{Outcome}.
\item \verb{Enrollment at entry} \emph{character} - Whether the students are full- or part-time.
\item \verb{Demographic Group} \emph{character} - A type of demographic group, like race or age.
\item \code{Detail} \emph{character} - A specific demographic group, like \code{Latine} or \code{18-24}.
\item \code{Cohort} \emph{integer} - The number of students that belong to this group.
\item \code{Outcome} \emph{numeric} - A quantitative \code{Measure} of some aspect of the school's performance.
}
}
\description{
Convert Higher Expectations' data layout to HERA's dashboard layout
}
