% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grads_times_credits.R
\name{read_grads_times_credits}
\alias{read_grads_times_credits}
\title{Import data on graduation times and credits earned across populations.}
\usage{
read_grads_times_credits(.file_name)
}
\arguments{
\item{.file_name}{the full path to the spreadsheet with data}
}
\value{
A tibble with seven fields:
\itemize{
\item \code{Degree/Certificate} \emph{character} - e.g., Certificate, Associate's, or Bachelor's.
\item \verb{Enrollment at entry} \emph{character} - Either Full- or Part-time.
\item \verb{Demographic Group} \emph{character} - What type of demographic group this row describes.
\item \code{Detail} \emph{character} - The specific demographic group that this row describes.
\item \code{Count} \emph{integer} - The number of students who were in the group.
\item \verb{Average credits to degree} \emph{numeric} - The average number of credits earned by students in the group.
\item \verb{Average time to degree} \emph{numeric} - The average number of years that students in this group took to graduate.
}
}
\description{
Import data on graduation times and credits earned across populations.
}
